# documentation/api/identity.openapi.yaml
openapi: 3.1.0
info:
  title: Zagrosi Identity API
  version: 0.1.0
  summary: Public identity, SSO, SCIM, and token-management API.
  description: >
    This specification documents the browser auth, OIDC, SAML, SCIM,
    session, personal access token, SCIM token, service token, and admin
    unlock surfaces shipped by the identity foundation.
  license:
    name: AGPL-3.0-or-later
    identifier: AGPL-3.0-or-later
...
servers:
  - url: https://api.zagrosi.example
    description: Production API gateway
security: []
...
paths:
...
  /v1/auth/saml/{org_slug}/start:
    get:
      tags: [SAML]
      operationId: startSaml
      summary: Start SP-initiated SAML Web SSO.
      parameters:
        - $ref: "#/components/parameters/OrgSlug"
      responses:
        "302":
          description: Redirect to the IdP.
          headers:
            Location:
              schema:
                type: string
                format: uri
  /v1/auth/saml/{org_slug}/acs:
    post:
      tags: [SAML]
      operationId: samlAcs
      summary: Consume a SAMLResponse POST binding.
      parameters:
        - $ref: "#/components/parameters/OrgSlug"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/SamlAcsRequest"
      responses:
        "302":
          description: Redirect after successful ACS.
          headers:
            Set-Cookie:
              $ref: "#/components/headers/SetCookie"
        "401":
          $ref: "#/components/responses/Problem401"
        "409":
          $ref: "#/components/responses/Problem409"
  /v1/auth/saml/{org_slug}/metadata.xml:
    get:
      tags: [SAML]
      operationId: samlMetadata
      summary: Return SAML SP metadata XML.
      parameters:
        - $ref: "#/components/parameters/OrgSlug"
      responses:
        "200":
          description: SAML metadata.
          content:
            application/samlmetadata+xml:
              schema:
                type: string
...
  /scim/v2/ServiceProviderConfig:
    get:
      tags: [SCIM]
      operationId: scimServiceProviderConfig
      summary: Return SCIM ServiceProviderConfig.
      security:
        - ScimBearer: []
      responses:
        "200":
          $ref: "#/components/responses/ScimResource"
  /scim/v2/ResourceTypes:
    get:
      tags: [SCIM]
      operationId: scimResourceTypes
      summary: Return SCIM resource type metadata.
      security:
        - ScimBearer: []
      responses:
        "200":
          $ref: "#/components/responses/ScimResource"
  /scim/v2/Schemas:
    get:
      tags: [SCIM]
      operationId: scimSchemas
      summary: Return SCIM schema metadata.
      security:
        - ScimBearer: []
      responses:
        "200":
          $ref: "#/components/responses/ScimResource"
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: __Host-zagrosi_sid
      description: Secure HttpOnly SameSite=Lax Path=/ session cookie.
    CsrfCookie:
      type: apiKey
      in: cookie
      name: __Host-zagrosi_csrf
      description: Double-submit CSRF cookie paired with `X-Zagrosi-CSRF`.
    OidcCsrfCookie:
      type: apiKey
      in: cookie
      name: __Host-zagrosi_oidc_csrf
      description: OIDC callback state cookie.
    BearerToken:
      type: http
      scheme: bearer
      bearerFormat: "pat_<43>, scim_<43>, svc_<43>"
    ScimBearer:
      type: http
      scheme: bearer
      bearerFormat: scim_<43>
  parameters:
...
    OrgSlug:
      name: org_slug
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-z0-9][a-z0-9-]{1,62}$"
...
  headers:
    SetCookie:
      schema:
        type: string
      description: >
        Session cookies use `__Host-zagrosi_sid; Path=/; Secure; HttpOnly;
        SameSite=Lax`. CSRF cookies use `__Host-zagrosi_csrf; Path=/; Secure;
        SameSite=Lax`.
...
    WWWAuthenticate:
      schema:
        type: string
      examples:
        bearer:
          value: Bearer
        scim:
          value: Bearer realm="SCIM"
    ETag:
      schema:
        type: string
      description: SCIM `meta.version` ETag.
...
  responses:
...
    Problem401:
      description: Unauthorized.
      headers:
        WWW-Authenticate:
          $ref: "#/components/headers/WWWAuthenticate"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    Problem409:
      description: Conflict.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    ScimResource:
      description: SCIM resource.
      headers:
        ETag:
          $ref: "#/components/headers/ETag"
      content:
        application/scim+json:
          schema:
            type: object
            additionalProperties: true
...
  schemas:
...
    SamlAcsRequest:
      type: object
      required: [SAMLResponse, RelayState]
      properties:
        SAMLResponse:
          type: string
        RelayState:
          type: string
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string